feature: add free functions on window close
add new headers for keys, macros and structs
