+++ include/keys.h
@@ -1,0 +1,40 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   keys.h                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: caide-so <caide-so@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/09 20:38:14 by caide-so          #+#    #+#             */
+/*   Updated: 2025/07/17 22:19:37 by caide-so         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef KEYS_H
+# define KEYS_H
+
+# define KEY_ESC	65307
+# define KEY_LEFT	65361
+# define KEY_UP		65362
+# define KEY_RIGHT	65363
+# define KEY_DOWN	65364
+
+# define KEY_PLUS	61
+# define KEY_MINUS	45
+# define KEY_SPACE	32
+
+# define KEY_Z		122
+# define KEY_X		120
+# define KEY_P		112
+# define KEY_I		105
+# define KEY_O		111
+# define KEY_L		108
+# define KEY_R		114
+# define KEY_A		97
+# define KEY_S		115
+# define KEY_D		100
+# define KEY_Q		113
+# define KEY_W		119
+# define KEY_E		101
+
+#endif
