+++ src/free/free_all.c
@@ -1,16 +1,34 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   free_config.c                                      :+:      :+:    :+:   */
+/*   free_all.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marcudos <marcudos@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: caide-so <caide-so@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/07/11 19:08:09 by marcudos          #+#    #+#             */
-/*   Updated: 2025/07/15 22:02:13 by marcudos         ###   ########.fr       */
+/*   Created: 2025/07/17 22:31:13 by caide-so          #+#    #+#             */
+/*   Updated: 2025/07/17 22:31:52 by caide-so         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/cub3d.h"
+
+void	free_game(t_game *game)
+{
+	if (!game)
+		return ;
+	if (game->img)
+		free(game->img);
+	if (game->cfg)
+		free_config(game->cfg);
+	if (game->win)
+		mlx_destroy_window(game->mlx, game->win);
+	if (game->mlx)
+	{
+		mlx_destroy_display(game->mlx);
+		free(game->mlx);
+	}
+	free(game);
+}
 
 void	free_config(t_cfg *cfg)
 {
@@ -27,4 +45,19 @@ void	free_config(t_cfg *cfg)
 	if (cfg)
 		free(cfg);
 	cfg = NULL;
+}
+
+void	free_map_copy(char **map, int height)
+{
+	int	i;
+
+	if (!map)
+		return ;
+	i = 0;
+	while (i < height)
+	{
+		free(map[i]);
+		i++;
+	}
+	free(map);
 }
