--- suppression.sup
@@ -1,147 +1,0 @@
-{
-   <MLX_CODAM>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:_dl_catch_exception
-   ...
-}
-{
- <MLX_CODAM>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:mlx_init
-   ...
-}
-{
- <MLX_CODAM>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:XrmGetStringDatabase
-   ...
-}
-{
- <MLX_CODAM>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:_XrmInitParseInfo
-   ...
-}
-{
- <MLX_CODAM>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:__tls_get_addr
-   ...
-}
-{
- <MLX_CODAM>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:__pthread_once_slow
-   ...
-}
-{
- <MLX_CODAM>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:glfwInit
-   ...
-}
-{
- <MLX_CODAM>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:_dl_init
-   ...
-}
-
-{
-   <MLX_CODAM_STRNCMP>
-   Memcheck:Addr8
-   fun:strncmp
-   fun:is_dst
-   fun:_dl_dst_count
-   fun:expand_dynamic_string_token
-   fun:fillin_rpath.isra.0
-   fun:decompose_rpath
-   fun:cache_rpath
-   fun:_dl_map_object
-}
-
-{
-   <MLX_CODAM_DLOPEN>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   fun:dl_open_worker
-   fun:_dl_open
-}
-
-{
-   <MLX_CODAM_SWR>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   fun:calloc
-   obj:/usr/lib/x86_64-linux-gnu/dri/swrast_dri.so
-}
-
-{
-   <MLX_CODAM_MLX>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   fun:mlx_init
-   fun:mlx_init_render
-   fun:mlx_compile_shader
-}
-
-{
-   MLX42_libdl_dlopen_related
-   Memcheck:Leak
-   match-leak-kinds: all
-   fun:malloc
-   ...
-   fun:dl_open_worker
-}
-
-{
-   MLX42_strncmp_invalid_read
-   Memcheck:Addr8
-   fun:strncmp
-   ...
-   fun:_dl_dst_substitute
-}
-
-{
-   MLX42_rtld_malloc_leak
-   Memcheck:Leak
-   fun:malloc
-   ...
-   fun:_dl_find_object_update
-}
-
-{
-   mlx42_dl_strncmp
-   Memcheck:Addr8
-   fun:strncmp
-   fun:is_dst
-   fun:_dl_dst_count
-   fun:expand_dynamic_string_token
-   fun:fillin_rpath.isra.0
-   fun:decompose_rpath
-   fun:cache_rpath
-   fun:cache_rpath
-   fun:_dl_map_object
-   fun:openaux
-   fun:_dl_catch_exception
-   fun:_dl_map_object_deps
-   fun:dl_open_worker_begin
-   fun:_dl_catch_exception
-   fun:dl_open_worker
-}
